+++ library/src/methods/strict/strict.ts
@@ -28,9 +28,11 @@ export function strict<TSchema extends ObjectSchema<any>>(
     _parse(input, info) {
       const result = schema._parse(input, info);
       return !result.issues &&
-        // Check length of input and output keys
-        Object.keys(input as object).length !==
-          Object.keys(result.output).length
+        // Check length of input and output values
+        Object.values(input as object).filter((value) => value !== undefined)
+          .length !==
+          Object.values(result.output).filter((value) => value !== undefined)
+            .length
         ? getIssues(info, 'object', 'strict', error || 'Invalid keys', input)
         : result;
     },